Rename common/index to common/math
