+++ Template3D/components/com_transform.ts
@@ -1,5 +1,5 @@
-import {Mat4, Quat, Vec3} from "../../common/index.js";
 import {create} from "../../common/mat4.js";
+import {Mat4, Quat, Vec3} from "../../common/math.js";
 import {Entity, Game} from "../game.js";
 import {World} from "../world.js";
 import {Has} from "./com_index.js";
