+++ common/math.ts
