+++ common/vec3.ts
@@ -1,5 +1,5 @@
-import {Mat4, Vec3} from "./index.js";
 import {get_translation} from "./mat4.js";
+import {Mat4, Vec3} from "./math.js";
 
 export function add(out: Vec3, a: Vec3, b: Vec3) {
     out[0] = a[0] + b[0];
