+++ src/components/com_camera.ts
@@ -1,5 +1,5 @@
-import {Mat4} from "../../common/index.js";
 import {create, perspective} from "../../common/mat4.js";
+import {Mat4} from "../../common/math.js";
 import {Entity, Game} from "../game.js";
 import {Has} from "./com_index.js";
