+++ src/components/com_render_shaded.ts
@@ -1,4 +1,4 @@
-import {Vec4} from "../../common/index.js";
+import {Vec4} from "../../common/math.js";
 import {
     GL_ARRAY_BUFFER,
     GL_ELEMENT_ARRAY_BUFFER,
