+++ src/components/com_transform2d.ts
@@ -1,5 +1,5 @@
-import {Mat2D, Rad, Vec2} from "../../common/index.js";
 import {create} from "../../common/mat2d.js";
+import {Mat2D, Rad, Vec2} from "../../common/math.js";
 import {Entity, Game} from "../game.js";
 import {Has} from "./com_index.js";
