+++ src/systems/sys_draw.ts
@@ -1,5 +1,5 @@
-import {Vec3} from "../../common/index.js";
 import {get_translation} from "../../common/mat4.js";
+import {Vec3} from "../../common/math.js";
 import {transform_point} from "../../common/vec3.js";
 import {DrawKind, DrawMarker} from "../components/com_draw.js";
 import {Has} from "../components/com_index.js";
